refactor: split models into files
